LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/try-tactic-basic-expected.txt


FAIL CSS Anchor Positioning: try-tactic assert_equals: offsetLeft expected 10 but got 99999
FAIL CSS Anchor Positioning: try-tactic 1 assert_equals: offsetLeft expected 10 but got 99999
FAIL CSS Anchor Positioning: try-tactic 2 assert_equals: offsetLeft expected 360 but got 99999
FAIL CSS Anchor Positioning: try-tactic 3 assert_equals: offsetLeft expected 360 but got 99999
FAIL CSS Anchor Positioning: try-tactic 4 assert_equals: offsetLeft expected 360 but got 99999
FAIL CSS Anchor Positioning: try-tactic 5 assert_equals: offsetLeft expected 20 but got 99999
FAIL CSS Anchor Positioning: try-tactic 6 assert_equals: offsetLeft expected 20 but got 99999
FAIL CSS Anchor Positioning: try-tactic 7 assert_equals: offsetLeft expected 20 but got 99999
FAIL CSS Anchor Positioning: try-tactic 8 assert_equals: offsetLeft expected 20 but got 99999
FAIL CSS Anchor Positioning: try-tactic 9 assert_equals: offsetLeft expected 20 but got 99999
FAIL CSS Anchor Positioning: try-tactic 10 assert_equals: offsetLeft expected 340 but got 99999
FAIL CSS Anchor Positioning: try-tactic 11 assert_equals: offsetLeft expected 340 but got 99999
FAIL CSS Anchor Positioning: try-tactic 12 assert_equals: offsetLeft expected 340 but got 99999
FAIL CSS Anchor Positioning: try-tactic 13 assert_equals: offsetLeft expected 340 but got 99999
FAIL CSS Anchor Positioning: try-tactic 14 assert_equals: offsetLeft expected 340 but got 99999
FAIL CSS Anchor Positioning: try-tactic 15 assert_equals: offsetLeft expected 340 but got 99999
